feat: add author check to delete question use case

+++ src/domain/forum/application/use-cases/delete-question.ts
@@ -2,6 +2,7 @@ import { QuestionRepository } from '@/domain/forum/application/repositories/ques
 
 interface DeleteQuestionUseCaseRequest {
   questionId: string
+  authorId: string
 }
 
 // eslint-disable-next-line @typescript-eslint/no-empty-object-type
@@ -11,11 +12,15 @@ export class DeleteQuestionUseCase {
   constructor(private readonly questionRepository: QuestionRepository) {}
   async execute({
     questionId,
+    authorId,
   }: DeleteQuestionUseCaseRequest): Promise<DeleteQuestionUseCaseResponse> {
-    const foundQuestion = this.questionRepository.findById(questionId)
+    const foundQuestion = await this.questionRepository.findById(questionId)
 
     if (!foundQuestion)
       throw new Error('No question was found with the given ID')
+
+    if (foundQuestion.authorId !== authorId)
+      throw new Error('User not allowed to delete this question')
 
     await this.questionRepository.deleteById(questionId)
